+++ packages/chaingraph-types/src/decorator/secret.decorator.ts
@@ -6,17 +6,17 @@
  * As of the Change Date specified in that file, in accordance with the Business Source License, use of this software will be governed by the Apache License, version 2.0.
  */
 
-import type { PortDecoratorOptions } from './port.decorator.types'
-import { Port } from './port.decorator'
-
-/**
- * Decorator to specify a type configuration for type "secret".
- */
-export function Secret(config?: Omit<PortDecoratorOptions<'secret'>, 'type' | 'defaultValue'>): PropertyDecorator {
-  return Port({
-    type: 'secret',
-    secretType: 'string',
-    defaultValue: undefined,
-    ...config,
-  })
-}
+// import type { PortDecoratorOptions } from './port.decorator.types'
+// import { Port } from './port.decorator'
+//
+// /**
+//  * Decorator to specify a type configuration for type "secret".
+//  */
+// export function Secret(config?: Omit<PortDecoratorOptions<'secret'>, 'type' | 'defaultValue'>): PropertyDecorator {
+//   return Port({
+//     type: 'secret',
+//     secretType: 'string',
+//     defaultValue: undefined,
+//     ...config,
+//   })
+// }