+++ packages/typescript-config/CHANGELOG.md
@@ -1,5 +1,35 @@
 # @badaitech/typescript-config
 
+## 0.1.6-dev.4
+
+### Patch Changes
+
+- One more dev bundle testing
+
+## 0.1.6-dev.3
+
+### Patch Changes
+
+- Next try
+
+## 0.1.6-dev.2
+
+### Patch Changes
+
+- adjust the frontend lib vite config
+
+## 0.1.6-dev.1
+
+### Patch Changes
+
+- remove rollupOptions external from the frontend lib package vite config
+
+## 0.1.6-dev.0
+
+### Patch Changes
+
+- Remove secret port type and all secret usages
+
 ## 0.1.5
 
 ### Patch Changes